+++ lib/quick_travel/booking.rb
@@ -156,6 +156,10 @@ module QuickTravel
       reserve(:packages, options)
     end
 
+    def delete_reservations
+      delete_and_validate("#{api_base}/#{@id}/reservations")
+    end
+
     # Delete a reservation
     #
     # Returns current booking object after deleting the reservation